Move getting canister methods down to query/update
